set Xint as i64

+++ src/cell.rs
@@ -3,7 +3,7 @@ use crate::state::State;
 
 pub type Xvec = rpds::Vector<Cell>;
 pub type XfnType = fn(&mut State) -> Xresult;
-pub type Xint = i128;
+pub type Xint = i64;
 pub type Xreal = f64;
 pub type Xanyrc = std::rc::Rc<std::cell::RefCell<dyn std::any::Any>>;
 pub type Xbitstr = crate::bitstring::Bitstring;
